Import shared constance from the functions directory.
Files under functions/src can only import modules from functions/src,
while components/* can import from functions/src as long as
the modules only use deps that work in node and browser.

This moves the constants into a constants-only module that can be imported
from both packages.
