+++ components/db/members.ts
@@ -1,8 +1,8 @@
 import { Timestamp } from "firebase/firestore"
 import { useMemo } from "react"
 import { useAsync } from "react-async-hook"
-import { currentGeneralCourt, loadDoc } from "./common"
-
+import { currentGeneralCourt } from "functions/src/shared"
+import {loadDoc} from './common'
 export type CommitteeReference = {
   CommitteeCode: string
   GeneralCourtNumber: number
